+++ custom-plugins/msupply-worldmap/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './DataPointLayer';
+export * from './GeoJsonLayer';
